globallimiter: Improve reliability
When redis-store goes down (during rollouts, for example) all upstream operations that request a rate limit token will fail immediately. This is not great.

Instead, we are trying an approach now that adds retries.
We attempt to get the rate limit up to 8 times, and wait between 50 and 250ms for a total of 400ms to 2000ms of additional wait time before failing now. This number is just guesswork, and will not fully cover a rollout, but keep delays low. We will test this on dotcom and see what a good number would be.

In addition, instead of failing we are now falling back to an in-memory limiter instead. The limiter uses either 1/s as it's internal rate, or siteconfig.defaultRateLimit, if configured.