Include the last sub-sitemap in the sitemap index (#56526)
The sitemap command generates sitemaps in chunks of 50,000 URLs, but wasn't including the final chunk in the sitemap index. Make sure it is included by combining sitemap creation and index insertion.

Update the README to specify the content-encoding on the uploaded files. Without this, Google Cloud serves these without content-encoding and user agents get garbage/have to guess the encoding.

# Test plan

Builds, but this panics at main when there are no LSIF indexes.