gitservice: use an error hook instead of a logger
src-cli doesn't use logger, so this changes the pattern to instead make
how we log injectable by the client. Currently the logger is just not
set in src-cli leading to a nil pointer panic. After this is merged a
follow up PR will go out for that.

Additionally this PR fixes a use for sourcegraph app which didn't set
the logger as well.

Test Plan: CI