Simplify RepoNotFoundPage (#56606)
Currently, the repo not found page, if viewed by a site-admin tries to see if a repo might be cloneable. Then, it renders a slightly different error message, based on that result.
If cloneable: As a site admin, you can add this repository to Sourcegraph to allow users to search and view it by connecting an external service.
If not cloneable: If this is a private repository, check that this site is configured with a token that has access to this repository. If this is a public repository, check that this repository is explicitly listed in an external service configuration.

Both are very similar to me, and the not cloneable message does not add a ton of value to me, in fact, it even seems misleading: It suggests that the tokens in external service configs would have been used to see if the repo is accessible by sourcegraph, but in fact we make unauthenticated requests always!

What this did:
It calls gitserver.IsRepoCloneable with the name set to the literal name of the repo from the url, for example github.com/sourcegraph/sourcegraph. Remember: github.com/sourcegraph/sourcegraph is not a repo known to the instance, otherwise we wouldn't render the RepoNotFoundPage. In gitserver, we try to get the RemoteURL of this repo, which calls gitserverRepos.GetByName(name). Again, this repo is not known to the instance, so this lookup has to fail.
We then fall back to using the VCSGitSyncer with https://${name}/.git and try to see if that's cloneable. With no token, no nothing, and always a VCSGitSyncer, never a perforce syncer or so. I don't think this check adds any significant value, and adds complexity and unauthenticated requests which could cause IP based abuse protection from github.com, so I'm proposing that we remove it.

The wonky logic where we try to construct some arbitrary URL unfortunately still has to remain, because of a dotcom code path that checks if a repo can be cloned (ie. is public) before enqueueing a clone job. But as a next step, we could simplify this endpoint and give it a more descriptive name, at least.