uploadstore: list with prefix (#56538)
We extend Store.List and allow the caller to filter the objects by
prefix.

Store.List isn't used anywhere, so it is fine to modify it
instead of adding a dedicated new method.

Why isn't Store.List use anywhere?

I added this method a while ago to support the effort to create a single
source of truth for embeddings. However, in end I decided for a
different approach. I left the method in place because I anticipated we
might use it soon for other purposes.

Test plan:
- updated unit tests
- manually tested with local blobstore and s3 bucket in eu-central-1