JetBrains: add button to change codebase context
Previously, there was no way for users to hard-code a specific
repository that should be used for the embeddings context. This PR fixes
the problem by making the repository name a button that opens a form to
change the codebase when you click on it. The name of the codebase is
validated with the API so help the user prevent writing a typo.