search: split up repos resolver into DB part and gitserver part
We want this behaviour for search jobs so we can only query the DB and
at a later stage query gitserver (which is slower). This also helps
break up that monster function a bit. This commit should not change any
behaviour.

Additionally we add some doc strings to explain behaviour.

Test Plan: go test