search: alert reposExist helper handles empty pages (#56707)
This was a likely very rare to trigger bug. The more likely bug to
trigger was it asserted err was nil, but you can find a repo but have
some revs missing (which would result in an error). So this changes
behaviour, but looking at the call sites, function name and
documentation this is now the correct behaviour.

Note: I plan on adding an API which only communicates with the DB and
avoids slower gitserver calls. That may be useful to use here in the
future since this helper is only used for suggestions.

Test Plan: CI