search: exhaustive RepositoryRevSpec -> RepositoryRevSpecs (#56755)
Previously we would create a worker job per repo@refspec. However, we
actually need the list of refspecs from a query to live together since
they influence each other. For example the negated glob can remove refs
from the final generated list of refs to search.

This commit keeps the string type and doesn't change anything in the DB.
However, it updates naming and documentation to reflect that it is a
list of refspecs.

I think there are database constraints around this, but I think once we
have stuff working in a nice way we should revisit things like the
schema and adjust it then.

Test Plan: CI