JetBrains: don't block forever on event thread (#56767)
Previously, we were sending a blocking network request on the IntelliJ event thread
while validating that a repository name exists on the Sourcegraph
instance in the dialog to select the Cody codebase context. If the
validation didn't return then the IDE was completely frozen. The only
wan to fix the problem was to force quit IntelliJ.

This PR fixes the problem by setting a timeout on the blocking requests
so that the user at least regains control over their IntelliJ instance.