user onboarding: Use server side configuration and improve admin experience (#56768)
This PR integrates the onboarding tour with the admin configuration, i.e. the onboarding tour now uses the content configuration from the server.

Additionally I made some changes to the admin page to improve UX:

- Show a simple of preview of the onboarding tour, which live updates as the config is edited
- Added separate schema validation (outside of Monaco) to
  - ensure that the preview always gets a valid config
  - ensure that the user cannot save an invalid schema (and break the tour for other users)
  - make errors in the config more visible
- Added a "Reset" button to restore the original (hardcoded) onboarding tour so that the user can always go back to a valid config.

It also fixes some issues overlooked in the previous PR (such as schema issues).

Note for the schema validation: I also could have added validation logic when loading the config for the real tour, but that would add a couple of dependencies to the "normal" pages that shouldn't be necessary. Preventing an invalid config from getting in in the first place seems better.

Note on the preview: In order to make the preview work I moved some logic into the tour context, do that those values can be overwritten easily for the preview. Specifically this affects the dynamic query generation, which we don't want to do for the preview.