Enable "Test connection" for Perforce code hosts (#56697)
* fill out `CheckConnection` and mark Perforce as able to check connections.

* Refactor the connection test to not rely on IsCloneable
Because connecting to Perforce requires using the `p4` CLI binary,
which is on `gitserver`, we're using the Perforce VCS Syncer to do the connection check.
Add a method to the VCS syncer for Perforce specifically for checking the connection
instead of hacking a solution that re-uses IsCloneable, which requires a valid depot.
Adding another func won't break composition with VCSSyncer.

* update comment on CheckConnection

* add a 10-second timeout for the connection test
10 seconds is somewhat arbitrary, but seems like enough time to allow
for various connection issues, whle not being too long for the frontend.

* Change from directly using the syncer
to making an RPC call to gitserver because the call to
CheckConnection comes fro mthe frontend, where `p4` is not available.

(cherry picked from commit b1fd4dab75f23230337e2b4a948ce9e43a882d9f)