source: make gitserver disk threshold configurable  (#56788)
The initial implementation of the gitserver disk threshold alert had a hard-coded value for the threshold before displaying an alert. It'll be nice for admins to be able to configure what this threshold should be.

This PR also moves the warning messages from being an alert to a status message, as shown below.