WIP enforce passing in a non-zero userID
This adjusts the search API for exhaustive to enforce explicit passing
in of a userID to ensure we always run searches as the correct user.
This is to prevent accidently running as another user and thus exposing
results from repos a user is not allowed to see.

Test Plan: CI