call it "Cody" not "Cody AI"
"Cody AI" was one way we referred to Cody in its first couple months to make it clear that it's an
AI product. That is no longer needed, and it sounds stilted/hype-y.

The one place where we do need to say "Cody AI" is when referring to Cody's name on the VS Code
extension marketplace. There, the name `Cody` by itself is taken, so we need `Cody AI` (or something
else that is not just `Cody`, and we're sticking with `Cody AI` for now).