always enable applySearchQuerySuggestionOnEnter
This makes it so that the previously on-by-default behavior of acting on search query suggestions when the user presses <kbd>Enter</kbd> is now the only behavior. The alternate behavior where <kbd>Enter</kbd> merely added the suggestion to the query is no longer available.

This was set to on by default on https://sourcegraph.com 9 months ago at https://github.com/sourcegraph/deploy-sourcegraph-cloud/pull/17546.