always enable applySearchQuerySuggestionOnEnter
This makes it so that the previously on-by-default behavior is now the only behavior. This only affects the old query input, not the new query input. The behavior is confusing to describe, but it basically means that as you type in the old query input, there is no autocomplete item selected until you press the down arrow, so pressing <kbd>Enter</kbd> does not confusingly accept an autocomplete item that you did not intend.

This was set to on by default on https://sourcegraph.com 9 months ago at https://github.com/sourcegraph/deploy-sourcegraph-cloud/pull/17546, and has been on by default in all other instances for 1 year.