use plain input instead of Monaco/CodeMirror for secondary query inputs
The following secondary query input fields now use a plain HTML `<input>` field instead of a Monaco or CodeMirror field with syntax highlighting, validation, and autocomplete:

- Code insight create/edit form
- Search context create/edit form
- Saved search create/edit form
- Code monitoring create/edit form

This is a minor product UX regression, and we should add back a CodeMirror input field in these places. But right now, they rely on an old and deprecated implementation, so removing this tech/product debt now is overall worth it.