JetBrains: add restarting logic for agent
Previously, if the agent failed to start for some reason then we didn't
try restarting it on autocomplete/chat requests. This PR changes the
behavior to add restarting logic when the user submits a chat message
AND when the user triggers autocomplete with the alt-backslash shortcut
(but not automatic triggers).

To further aid in troubleshooting agent startup issues, this PR also:

- Redirects agent stderr into idea.log so that we capture unexpected
  errors starting the process (previously, stderr was lost).
- Adds a new GitHub issue template so that we can link to
  it in the error message when the agent is not running.