database/telemetryexport: store messages exactly as it will be exported (#57028)
This change makes it so that the contents of telemetry_events_export_queue are exactly what we export, by having sensitive field redaction occur up-front as we insert the events into the database queue.

This has benefits for #57027 as what you see in the telemetry_events_export_queue is now the exact payload that leaves a Sourcegraph instance.

This is not yet GA or enabled outside of internal instances, so there is no concern for any already-queued events.

(cherry picked from commit ef6f0cce450aff7b08a45be74d3f00a65445cf31)