actor: propagate anonymous UID from 'X-Sourcegraph-Actor-Anonymous-UID' (#57056)
Cookie anonymous UID continues to take precedence, but this is helpful for clients that don't use cookies to track anonymous UID (i.e. VSCode), and especially relevant in the context of our new telemetry mutations, which now use real actors to determine the user, instead of asking clients to tell us who the user is.