user onboarding: Add onboarding setup flow (#57014)
This PR adds the onboarding setup flow for the onboarding tour.

If a user has less than X (currently 5) queries and hasn't gone through the setup flow they will see the setup dialog when they open a page that would normally show the tour (search homepage, search results, file page). The user can either answer the questions or skip the whole setup. If they skip the setup the tour will still be shown but only show tasks that don't depend on this information.

## Implementation details

I didn't want to completely reinvent the wheel but also didn't have to the time to build a truly reusable autocompletion system for repositories and languages. Therefor I moved some code around to make the caching logic used by the new query input reusable (now CachedAsyncCompletionSource), which is now used by the query input as well as the repository input. The language autocompletion is all synchronous using fzf. (I think it makes sense in the long run to make autocompletion "sources" more reusable)

The user provided values are validated (not empty, repository exists).

Exposing the PopoverContentProps on ComboboxPopover is necessary to for controlling (and fixing) the position of the popover.

Co-authored-by: numbers88s <cesar.r.jimenez@gmail.com>