search jobs: compute aggregate state in db (#56957)
This PR moves the logic for computing the aggregate state from the resolver to the store.

Why?

To filter and sort by the aggregate state and make use of the paging helpers, we have to return the aggregate state from db instead of computing it in the resolver. 

Note:

There is a bit of awkwardness because only `ListSearchJobs` computes the aggregate state, but not `GetSearchJob`. I don't want to set `AggState` in `GetSearchJob` because it is fairly expensive to calculate, and we call `GetSearchJob` a lot as part of the security checks, so I wanted to keep it as lean as possible. 

In the future we should have a dedicated call to the db for security checks and use `GetSearchJob` only when we create a search job. Then we can calculate the aggregate state on GetSearchJob, too, and remove the special casing in the code. 

Test plan:
- New unit test
- manual testing: I ran a couple of search jobs locally to confirm whether the aggregate state properly reflect the state of the underlying jobs.