gitserver: Simplify getRemoteURLFunc
This code path was required for cloud v1 because we would have private repos alongside public repos, and cloud default services would usually not be able to pull those.
Since we don't have private repos anymore, we can skip this check.
Worst case, we will have a repo marked as private in the DB and we won't attempt to update it - which would fail anyways. But now with a no sources for X error instead.