gitserver: Don't attempt to clone repos without auth when DB fails
We have a code path for dotcom where we try to see if a repo is cloneable if it doesn't exist in the DB yet.
However, we assume that any error from the getRemoteURL means that the repo is not existing and we should try to clone the repo without auth from a madeup URL.
This could cause unauthed clone attempts to repos that actually DO exist in the instance, so added an additional check here if the error is actually a "repo isn't found" error.