repoupdater: Break dependency on cmd/gitserver
The repoupdater code would import from cmd/gitserver to get the perforce syncer. Also, it required a p4 binary in the service to list depos while syncing.
Since we want to be able to move pieces around, and p4 is not installed to all containers by default, we have to do this to get the external service syncer into worker.
Instead, we now have a dedicated endpoint that gitserver exposes that we can use to make this check.
This also fixes circular import issues I ran into while trying other refactorings, because gitserver could not import from internal/repos.