perforce: Make connection check a proper RPC call
This makes checking the connectivity to perforce a proper RPC call, these are easier to test in isolation and it doesn't depend on us being able to run the users command - instead we use the p4testwithtrust utility.
This also makes it so that we don't start a reader that we never consume.