gitserver: Make gitserver code internal
This puts the code of gitserver into a package called internal, to prevent imports from it from other binaries. This is in line with what our other services do, and prevents us from accidentally importing from gitserver, which could have undesirable side-effects like metrics registered that are not used and so forth.