search: return status forbidden on download exhaustive logs (#57099)
Previously we would always write a header line before actually checking if we
can view a job. This automatically lead to status 200 and us ignoring the
later error. This adjusts the service call to return before any writes.

Test Plan: a test was added which ensures we don't write anything in case of
auth failure.