gitserver: Introduce subpackages
The main goal of this PR is to get the vcssyncers into a separate package. There's more cleanup we can do after wrt naming, but this is a mostly mechanical change.
Moving this code into a separate package also required being able to access some other functions in multiple packages, so I created an
- executil package with helpers around exec.Command
- gitserverfs with helpers around everything disk-related in ReposDir