web: Add developer settings dialog (#57135)
This PR adds a "developer settings dialog" that is only available when Sourcegraph is run locally in dev mode, and provides easy access to feature flags and temporary settings. This can also be useful for people testing a PR.

QOL features:

- Filter flags/settings by status and name
- Indicator whether overrides exist and how many there are
- Save/restore dialog state to/from local storage

We can use this dialog to add more development specific settings in the future.

## Changes for feature flags

We already have a way to locally override feature flags and the dialog makes use of the same system. The only change I had to make was converting the type to a runtime value so that I can list all flags.

## Changes for temporary settings

Similar to feature flags I had to add a runtime representation of available temporary settings, but since they have more complex types I had to go about a different way (maybe someone has a better idea though). This requires the setting names to be listed twice, but I think that's OK.

Temporary settings didn't have a system for local overrides. I added one by wrapping an existing backend with a new override backend. Once a feature flag is overridden with setTemporarySettingOverride, the backend will return the overriden value instead. Similarly it will write an overridden setting to local storage and not to the real backend. In order to notify/update all subscribers when a value was overridden I added a global observer which the override backend listens to.