search: exhaustive handlers validate input before writing CSV (#57168)
This commit makes it such that both CSV handlers for exhaustive search
validate the job ID before setting any CSV headers or writing data. This
makes it possible to distingiush between errors that happen before we
write out CSVs and during.

Another benefit is we now call WriteHeader, which avoids long delays
between clicking the download button and the download starting. Before
until go decides to flush there may be several seconds before the
headers are sent.

To achieve this change we adjust the service interface to instead return
an io.WriterTo. A more generic API would return an io.Reader (involving
the use of an io.Pipe). However, we can specialize to io.WriterTo since
that is exactly the interface we want to use (and is an underlying
optimization used by io.Copy)

Additionally we add instrumentation to the logs service as well as
making the instrumentation able to distingiush between the fast
bits (speaking to our store) and the slow bits (writing to client).

Test Plan: I have a lot of confidence in our go tests to catch bugs
here. However, we do not have as rock solid automated tests for these
handlers. So did manual testing on the happy case. I am confident enough
to wait until a few commits later to do more intensive manual tests.