search: validate query before creating exhaustive job
Right now we don't have a good way to give feedback if a worker fails
due to the query not being passed. Right now we validate in the
frontend, but the validation isn't always aligned so it is possible to
hit this path. This commits adds a step in the job creation service
which first sees if the query parses.

Test Plan: go test should be sufficient.