sign in page: add return to query to auth provider button URLS
Each of the buttons for the 3rd party auth providers has a callback link that looks like:

http://sourcegraph.example.com/.auth/openidconnect/login?pc=fooBar

Note that this link doesn't contain a returnTo redirect parameter, which is what the logic in 463c2a4b7508e83eb19ad6ae66cc2d9dfff6444c expects. Since the returnTo parameter is missing, the user is navagated back to the root homepage.

This PR adds a workaround by manually adding the returnTo query parameter if it's present on the URL the user is using when viewing the sign-in page.