Embeddings: tolerate some failure when generating embeddings (#57224)
We regularly run into a problem where a customer with a large repo tries to embed that repo, gets part way, but then flakiness from OpenAI, their network, or whatever else causes the whole embeddings job to fail. This is very frustrating on the side of the customer, and is the source of a lot of delays when getting customers up and running with Cody. We already retry in the HTTP layer, but sometimes this isn't enough, especially since failures seem to be correlated with the content being sent.

Instead, this PR allows some requests to fail. We already allow some chunks to fail (which is the most common failure case where openAI just returns a null vector), but this extends that setting to now also include full request failures.

To avoid jobs succeeding when there is a configuration error, some persistent connection error, or the like, we allow some chunks to fail, but if more than 10% of the chunks fail to embed, we fail the whole job. This number is arbitrary, but it probably doesn't matter too much as long as it's greater than zero and less than 100.