sg: Install p4 cli for local dev
Since we are working with perforce a bunch and have test suites that need the CLI that are currently skipped locally, I think it makes sense to add the p4 CLI here.
The installation for ubuntu seems more involved and requires adding package sources and so forth, which I'm not sure we want to do. If you have a good idea on how to best implement this for linux as well, let me know.