search: exhaustive panic on uploadstore error (#57300)
If we fail to get from the upload store it would result in a panic due
to trying to close the reader. However, the only practical way to
trigger this path is to cancel the download request, which means the
user luckily wouldn't see this panic.

Test Plan: go test to ensure nothing regresses further. Otherwise I am
relying on just code review.