auth: return 401 for invalid auth headers on DotCom (#57160)
In our previous setup, when a request with incorrect authentication headers reached DotCom, it was mistakenly treated with a 429 status code due to how we handle rate limits for anonymous requests. DotCom mode-specific behaviour was identified: the system keeps forwarding the wrong header until it reaches the [middleware that checks if we allow anonymous access](https://sourcegraph.com/github.com/sourcegraph/sourcegraph/-/blob/cmd/frontend/auth/non_public.go?L23-37).
This middleware passes the check as it's configured to identify if we are operating in [DotCom mode](https://sourcegraph.com/github.com/sourcegraph/sourcegraph/-/blob/internal/conf/auth.go?L28-33).

This commit makes sure that when in DotCom mode and we have an invalid header, we return a 401 to prevent us from treating users that supply (invalid) authorization headers as anonymous users.