gitserver: Make gitserver code internal (#57122)
* gitserver: Make gitserver code internal

This puts the code of gitserver into a package called internal, to prevent imports from it from other binaries. This is in line with what our other services do, and prevents us from accidentally importing from gitserver, which could have undesirable side-effects like metrics registered that are not used and so forth.

* Fixup bad imports in test packages

The integration_tests package and a few of the client tests needed to be moved into the gitserver package - this doesn't feel wrong to me as when they do test the server code, they really are an integration test and not just an isolated client test anymore.