dotcom: Don't depend on gitserver for repo.add (#57311)
The IsRepoCloneable had two hats before:
- Check if the given repo that Sourcegraph knows about is cloneable; this is used for both dotcom and on-prem
- Check if an unknown repo name might be accessible to sourcegraph without any auth provided; this is solely used on dotcom to check if we should lazily add a new repo when visited

This code was confusing to me and nothing prevented it from technically running for on-prem as well, when a repo is lost in between a request for isCloneable coming in and it hitting gitserver - we then might just talk to a random upstream with no auth.

Up until recently, we did that for all repos because we were missing the errcode.IsNotFound check - meaning during DB outages we might have hit external hosts unexpectedly as well.

This also wouldn't have worked for non-git repos, because we just assumed all repos we see are git repos. For example, if we ever added perforce here this wouldn't work.

This PR makes the isRepoCloneable endpoints sole purpose to check if an existing repo can be cloned, using it's actual VCSSyncer implementation.

Instead of reaching out to gitserver, frontend simply checks if the git API endpoint is reachable, not requiring another roundtrip to an internal service, and making it work while gitserver is rolling out as well, as a nice side-effect.

I tested this new logic with both GitHub.com and GitLab.com and it seems to work reliably.