Allow to edit code host config temporarily in local dev by default
I often want to quickly test a config change to a code host config, but then realize I forgot to set this env var manually again and then need to restart the entire stack.
I don't think disallowing edits is a great default, so instead of adding it to my override file I'm proposing to add it to the general configuration.

What do people think?