search jobs: add state aggregation to GetSearchJob (#57383)
This removes the awkwardness introduced [here](#56957) between `Get` and `List` for exhaustive search. The problem was that we used `Get` to retrieve the job AND for security checks. This PR decouples these two concerns.

- We add a dedicated method to the store for security checks and update all call sites
- The remaining calls to Get are legit calls to retrieve the job
- We update `Get` to include the aggregated state, just like `List`.

Note:

This doesn't mean we shouldn't investigate a more elegant solution for state aggregation long term, but at least the complexity is now confined to the store and doesn't leak to the API anymore.

## Test plan
This is a refactor, so relying on CI