graphqlbackend: remove unused repositoryConnectionResolver (#57416)
This resolver is not actually used. I think it could of been removed a
while ago as per Naman's comment about it being unused. The interface
for it needs to remain, but the actual implementation here can be
removed.

Test Plan: go test. If that is happy we know the code is unused.