database: remove ReposListOptions UserID and OrgID
There are no more uses of these options and we want to remove the
columns from postgres at some point. Additionally we correctly enforced
and documented that SearchContextID is mutually exclusive with
ExternalServiceIDs. I looked at usages of ExternalServiceIDs and it will
never be set at the same time as SearchContextID (so no lurking bugs
here).

Test Plan: go test