search: do not specify user or org id from search contexts (#57420)
We don't have repos associated with user or org anymore. I was thinking
stuff would go wrong since we were setting these values, but if you read
the database code if SearchContextID is set we ignore the UserID and
OrgID.

This change is part of removing UserID and OrgID from the repos API.

Test Plan: go test. Additionally manually inspected some tables on
dotcom to understand what data we had for these fields.