repoupdater: Remove endpoints for external services
These three endpoints used to live in repo-updater for a single reason only: Because we wanted to stick to the rate limits.
However, now that the rate limiters are shared across services through redis, this no longer holds.

This PR does the following:
- Split the functionality of `SyncExternalService` into the two things it was used for: Triggering a sync, and validating the service is valid.
- Move the `ValidateConnection`, `DiscoverRepos`, `ListNamespaces` functions from RPC to local function calls in the frontend.
- Remove those endpoints from repoupdater effective immediately

This IS technically a breaking change. However, I think for a few reasons it is fine to stop repo updater from implementing these immediately, in favor of velocity and getting to less responsibilities of this service quicker:
- The ListNamespaces and DiscoverRepos function is only invoked in the setup wizard on user request. When a user is updating their systems at the same time as they are initially configuring them, I want to assume that other errors could just as well happen, like service down and so forth.
- The SyncExternalService endpoint is called on user request. Failure to do so will not cause the system to break. We regularly schedule services anyways. Easy to recover from IMO.