remove codicons from client/vscode
They were used in very minor ways (as button icons when there was already accompanying text). Their use significantly complicated the build process and the code. There is probably an easier way to use them that we could figure out given some time/research, but they are not a priority.