clean up client/vscode usage of icons and monaco (#57446)
* remove monaco remnants from client/vscode

The Sourcegraph VS Code extension (`client/vscode`) previously used the Monaco editor for the search input, as did the rest of Sourcegraph's UI. For almost a year, however, it has used CodeMirror. Thus we can remove the remnants of Monaco support in the VS Code extension.

* remove codicons from client/vscode

They were used in very minor ways (as button icons when there was already accompanying text). Their use significantly complicated the build process and the code. There is probably an easier way to use them that we could figure out given some time/research, but they are not a priority.